requirements.txt
amqp==2.1.1
billiard==3.5.0.2
celery==4.0.0
certifi==2016.9.26
contextlib2==0.5.4
Django==1.10.3
django-constance==1.3.3
django-picklefield==0.3.2
django-rest-framework-mongoengine==3.3.0
djangorestframework==3.5.3
ethereum-abi-utils==0.2.2
gevent==1.1.2
geventhttpclient==1.3.1
greenlet==0.4.10
kombu==4.0.0
mock==2.0.0
mongoengine==0.10.6
pbr==1.10.0
psycopg2==2.6.2
py==1.4.31
pylru==1.0.9
pymongo==3.3.1
pysha3==0.3
pytz==2016.7
raven==5.31.0
redis==2.10.5
rlp==0.4.6
sentinels==1.0.0
six==1.10.0
vine==1.1.3
web3==3.2.0
wheel==0.24.0
factory-boy==2.7.0
django-ipware==1.1.6